// temp_old_version/JB-Portal-dev-2/src/app/shared/interface/theme.interface.ts
export interface IHomeBanner {
  background_image?: string;
  status: boolean;
  banners: IBanners[];
}

export interface IBanners {
  title?: string;
  description?: string;
  tag?: string;
  offer?: string;
  button_text: string;
  redirect_link: IRedirectLink;
  image_url: string;
  status: boolean;
}
export interface IRedirectLink {
  link_type: string;
  link: string | number;
  product_ids: number;
}

export interface IFeaturedBannersFashionOne {
  banner_1?: IFeaturedBanner;
  banner_2: IFeaturedBanner;
}
export interface IFeaturedBanner {
  redirect_link?: IRedirectLink;
  image_url: string;
  tag?: string;
  title?: string;
  sub_title?: string;
  status: boolean;
}

export interface IProductList {
  tag?: string;
  title: string;
  sub_title?: string;
  description?: string;
  image_url?: string;
  product_ids: number[];
  status?: boolean;
}
// ...
export interface IService {
  status: boolean;
  banners: IBanners[];
}

export interface ISocialMedia {
  title?: string;
  status: boolean;
  banners: IBanners[];
}

export interface IBrands {
  brand_ids: number[];
  status: boolean;
}
// ...
export interface IBlogSection {
  tag?: string;
  title: string;
  sub_title?: string;
  description?: string;
  status: boolean;
  blog_ids: number[];
}
// ...
export interface IProductCategoryFashionFour {
  title?: string;
  category_ids: number[];
  status: boolean;
}
// ...
export interface IHomeBannerFashionFive {
  status: boolean;
  image_url: string;
  redirect_link: IRedirectLink;
}
// ...
export interface IVegetablesFour {
  id: number;
  content: IVegetablesFourContent;
  slug: string;
}

export interface IVegetablesFourContent {
  home_banner: IHomeBannerFashionFive;
  categories: IProductCategoryFashionFour;
  offer_banner_1: IHomeBanner;
  products_list_1: IProductList;
  products_list_2: IProductList;
  offer_banner_2: IHomeBanner;
  products_list_3: IProductList;
  services: IService;
  products_list_4: IProductList;
  featured_blogs: IBlogSection;
  brand: IBrands;
  products_ids: number[];
}